sindicatos_final/requirements_deploy.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
beautifulsoup4==4.13.4
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
SQLAlchemy==2.0.23
alembic==1.13.1
PyPDF2==3.0.1
pandas==2.1.4
openpyxl==3.1.2
Pillow==10.1.0
google-generativeai==0.3.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
numpy==1.26.0
pytz==2023.3.post1
six==1.16.0
urllib3==2.0.5
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
python-dateutil==2.8.2
tzdata==2023.3
scikit-learn==1.3.0
nltk==3.8.1
gunicorn==21.2.0
